+++ lib/danger/ci_source/semaphore.rb
@@ -4,7 +4,11 @@ module Danger
   module CISource
     class Semaphore < CI
       def self.validates?(env)
-        return !env["SEMAPHORE"].nil?
+        return false unless env["SEMAPHORE"]
+        return false unless env["SEMAPHORE_REPO_SLUG"]
+        return false unless env["PULL_REQUEST_NUMBER"].to_i > 0
+
+        return true
       end
 
       def supported_request_sources
@@ -13,9 +17,8 @@ module Danger
 
       def initialize(env)
         self.repo_slug = env["SEMAPHORE_REPO_SLUG"]
-        if env["PULL_REQUEST_NUMBER"].to_i > 0
-          self.pull_request_id = env["PULL_REQUEST_NUMBER"]
-        end
+        self.pull_request_id = env["PULL_REQUEST_NUMBER"]
+        self.repo_url = GitRepo.new.origins # Semaphore doesn't provide a repo url env variable :/
       end
     end
   end