+++ spec/lib/danger/ci_sources/circle_spec.rb
@@ -4,27 +4,36 @@ describe Danger::CISource::CircleCI do
   legit_pr = "https://github.com/orta/thing/pulls/45"
   not_legit_pr = "https://github.com/orta"
 
-  it 'validates when circle env var is found and it has a real PR url' do
-    env = { "CIRCLE_BUILD_NUM" => "true", "CI_PULL_REQUEST" => legit_pr }
+  it 'validates when circle all env vars are set' do
+    env = { "CIRCLE_BUILD_NUM" => "true",
+            "CI_PULL_REQUEST" => legit_pr,
+            "CIRCLE_PROJECT_USERNAME" => "orta",
+            "CIRCLE_PROJECT_REPONAME" => "thing" }
     expect(Danger::CISource::CircleCI.validates?(env)).to be true
   end
 
   it 'validates when circle env var is found and it has a bad PR url' do
-    env = { "CIRCLE_BUILD_NUM" => "true", "CI_PULL_REQUEST" => not_legit_pr }
-    expect(Danger::CISource::CircleCI.validates?(env)).to be true
-  end
-
-  it 'validates when circle env var is found and it has no PR url' do
     env = { "CIRCLE_BUILD_NUM" => "true",
-      "CIRCLE_PROJECT_USERNAME" => "orta",
-      "CIRCLE_PROJECT_REPONAME" => "thing" }
+            "CI_PULL_REQUEST" => not_legit_pr,
+            "CIRCLE_PROJECT_USERNAME" => "orta",
+            "CIRCLE_PROJECT_REPONAME" => "thing" }
     expect(Danger::CISource::CircleCI.validates?(env)).to be true
   end
 
   it 'doesnt get a PR id when it has a bad PR url' do
-    env = { "CIRCLE_BUILD_NUM" => "true", "CI_PULL_REQUEST" => not_legit_pr }
+    env = { "CIRCLE_BUILD_NUM" => "true",
+            "CI_PULL_REQUEST" => not_legit_pr,
+            "CIRCLE_PROJECT_USERNAME" => "orta",
+            "CIRCLE_PROJECT_REPONAME" => "thing" }
     t = Danger::CISource::CircleCI.new(env)
     expect(t.pull_request_id).to be nil
+  end
+
+  it 'doesnt validate when circle env var is found and it has no PR url' do
+    env = { "CIRCLE_BUILD_NUM" => "true",
+            "CIRCLE_PROJECT_USERNAME" => "orta",
+            "CIRCLE_PROJECT_REPONAME" => "thing" }
+    expect(Danger::CISource::CircleCI.validates?(env)).to be false
   end
 
   it 'doesnt validate when circle ci is not found' do