+++ spec/lib/danger/danger_core/environment_manager_spec.rb
@@ -4,7 +4,7 @@ describe Danger::EnvironmentManager do
   it 'raises without enough info in the ENV' do
     expect do
       Danger::EnvironmentManager.new({ "KEY" => "VALUE" })
-    end.to raise_error("Could not find a CI source".red)
+    end.to raise_error("Could not find a valid pull request within the known CI sources".red)
   end
 
   it 'stores travis in the source' do
@@ -16,19 +16,26 @@ describe Danger::EnvironmentManager do
 
   it 'stores circle in the source' do
     number = 800
-    env = { "CIRCLE_BUILD_NUM" => "true", "CI_PULL_REQUEST" => "https://github.com/artsy/eigen/pull/#{number}" }
+    env = { "CIRCLE_BUILD_NUM" => "true",
+            "CI_PULL_REQUEST" => "https://github.com/artsy/eigen/pull/#{number}",
+            "CIRCLE_PROJECT_USERNAME" => "orta",
+            "CIRCLE_PROJECT_REPONAME" => "thing" }
     e = Danger::EnvironmentManager.new(env)
     expect(e.ci_source.pull_request_id).to eq(number.to_s)
   end
 
   it 'creates a GitHub attr' do
-    env = { "HAS_JOSH_K_SEAL_OF_APPROVAL" => "true", "TRAVIS_REPO_SLUG" => "KrauseFx/fastlane", "TRAVIS_PULL_REQUEST" => 123.to_s }
+    env = { "HAS_JOSH_K_SEAL_OF_APPROVAL" => "true",
+            "TRAVIS_REPO_SLUG" => "KrauseFx/fastlane",
+            "TRAVIS_PULL_REQUEST" => 123.to_s }
     e = Danger::EnvironmentManager.new(env)
     expect(e.request_source).to be_truthy
   end
 
   it 'skips push runs and only runs for pull requests' do
-    env = { "TRAVIS_REPO_SLUG" => "orta/danger", "TRAVIS_PULL_REQUEST" => "false", "HAS_JOSH_K_SEAL_OF_APPROVAL" => "1" }
+    env = { "TRAVIS_REPO_SLUG" => "orta/danger",
+            "TRAVIS_PULL_REQUEST" => "false",
+            "HAS_JOSH_K_SEAL_OF_APPROVAL" => "1" }
     e = Danger::EnvironmentManager.new(env)
     expect(e.ci_source).to eq(nil)
   end