+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Support multiple Danger instances with `--dangerId` - marcelofabri
 * Add base request source so services other than GitHub could be used with Danger. - justMaku
+* Don't validate CI sources that don't expose all required environment variables.
 
 ## 0.8.1